[Refactor] Moved Sequencer class object into Logic_Tests class. This will enable Logic_Tests to do its own randomized value testing internally.
